requirements.txt
# ==============================================
# Swing Trading Copilot - Production Dependencies
# ==============================================

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
flask-caching==2.1.0
flask-sqlalchemy==3.1.1

# Production Server
gunicorn==21.2.0

# Database Drivers
psycopg2-binary==2.9.9

# API Clients
alpaca-py==0.14.0
polygon-api-client==1.13.1
requests==2.31.0
openai==1.3.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
pandas-ta>=0.3.14b

# Environment Variables
python-dotenv==1.0.0

# Payment Processing
stripe==7.8.0
